Fix comment and type annotation for filter function
The filter need to be moved to a separate file because else we get
cyclic includes.
