+++ src/gcovr/formats/cobertura/read.py
@@ -22,20 +22,20 @@ import os
 from glob import glob
 from lxml import etree  # nosec # We only write XML files
 
-from ...options import Options
 from ...coverage import (
     BranchCoverage,
     CoverageContainer,
     FileCoverage,
     LineCoverage,
 )
+from ...filter import is_file_excluded
 from ...merging import (
     get_merge_mode_from_options,
     insert_branch_coverage,
     insert_file_coverage,
     insert_line_coverage,
 )
-from ...utils import is_file_excluded
+from ...options import Options
 
 LOGGER = logging.getLogger("gcovr")
 
@@ -72,7 +72,12 @@ def read_report(options: Options) -> CoverageContainer:
         except Exception as e:
             raise RuntimeError(f"Bad --cobertura-add-tracefile option.\n{e}") from None
 
-        source_dir = str(root.find("./sources/source").text)
+        source_elem = root.find("./sources/source")
+        if source_elem is None:
+            raise AssertionError(
+                f"No source directory defined in file {data_source_filename}"
+            )
+        source_dir = str(source_elem.text)
 
         gcovr_file: etree._Element
         for gcovr_file in root.xpath("./packages//class"):  # type: ignore [assignment, union-attr]
@@ -83,13 +88,11 @@ def read_report(options: Options) -> CoverageContainer:
                 )
                 continue
 
-            filename = str(os.path.normpath(filename))
+            filename = str(os.path.normpath(os.path.join(source_dir, filename)))
             if is_file_excluded(filename, options.filter, options.exclude):
                 continue
 
-            filecov = FileCoverage(
-                os.path.join(source_dir, filename), data_source_filename
-            )
+            filecov = FileCoverage(filename, data_source_filename)
             merge_options = get_merge_mode_from_options(options)
             xml_line: etree._Element
             for xml_line in gcovr_file.xpath("./lines//line"):  # type: ignore [assignment, union-attr]
