+++ src/gcovr/formats/cobertura/write.py
@@ -17,7 +17,6 @@
 #
 # ****************************************************************************
 
-from __future__ import annotations
 from dataclasses import dataclass
 import os
 from lxml import etree  # nosec # We only write XML files
